+++ modules/net-vpc-peering/variables.tf
@@ -15,7 +15,7 @@
  */
 
 variable "prefix" {
-  description = "Name prefix for the network peerings"
+  description = "Name prefix for the network peerings."
   type        = string
   default     = "network-peering"
 }
@@ -46,4 +46,10 @@ variable "module_depends_on" {
   description = "List of modules or resources this module depends on."
   type        = list
   default     = []
+}
+
+variable "peer_create_peering" {
+  description = "Create the peering on the remote side. If false, only the peering from this network to the remote network is created."
+  type        = bool
+  default     = true
 }